fix: Change dependencies to resolve 'circular reference'
Normally I'll just stick the Spring Boot version to the same with the
other projects, but since this is just a generation service, it does
not really matter. Basically, if you use the default pom.xml provided
by renren-generator, you wouldn't be able to start the application
correctly, especially if you run this application in the debug mode,
something like '.. Bean already defined with the same name ..' and
'.. Is there an unresolvable circular reference ..'. If you try to
modify the DAOs and the mapper XML files, it just won't work (just
like I did). In the end, I got this project working by changing the
versions of Spring Boot and PageHelper.